requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# LLM Client (OpenAI-compatible SDK for Grok)
openai==1.58.1

# Async
aiohttp==3.11.11
aiofiles==24.1.0
tenacity==9.0.0
asyncio-throttle==1.0.2

# Parsing
tree-sitter==0.25.2
tree-sitter-python==0.23.6
tree-sitter-javascript==0.23.1
tree-sitter-typescript==0.23.2
tree-sitter-go==0.23.4
tree-sitter-java==0.23.5
tree-sitter-rust==0.23.2
tree-sitter-c==0.23.4
tree-sitter-cpp==0.23.4

# Vector Store
chromadb==0.6.3

# Git
gitpython==3.1.44

# Utilities
uuid6==2024.7.10
orjson==3.10.12
httpx==0.28.1